Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-float main()
+int main()
 {
     float a,b,product;
     scanf("%f",&a,&b);
